==> notebooks/01-languge_modeling.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import os
os.chdir('/content/drive/MyDrive/innovae-revision/innovae-adavae')

!pip install transformers
!pip install tensorboardX

cd adavae/src

!python adaVAE.py \
    --batch-sizes 40 --dataset patent_claim \
    --max_length 400 --pre_enc_iter start \
    --add_attn --beta_0 1 \
    --fb 1 --latent_gen latent_attn \
    --adapter_size 128 --iterations 26272 \
    --weighted_sample \
    --latent_size 512 \
    --encoder_n_layer 8 \
    --decoder_n_layer 12 \
    --adapter_init bert \
    --attn_mode none  \
    --kl_rate 0.50
